+++ Emulator/Cartridges/CustomCartridges/Isepic.h
@@ -87,8 +87,8 @@ private:
     //
     
     bool hasSwitch() const override { return true; }
-    const string getSwitchDescription(i8 pos) const override;
-    void setSwitch(i8 pos) override;
+    const string getSwitchDescription(isize pos) const override;
+    void setSwitch(isize pos) override;
     bool switchInOffPosition() const { return switchIsLeft(); }
     bool switchInOnPosition() const { return switchIsRight(); }
